+++ src/lib.rs
@@ -139,13 +139,9 @@ extern crate rusttype;
 extern crate serde_derive;
 extern crate serde;
 
-#[cfg(feature = "conrod")]
-pub extern crate conrod_core as conrod;
 #[cfg(not(target_arch = "wasm32"))]
 extern crate glutin;
 extern crate instant;
-#[cfg(feature = "conrod")]
-pub use conrod::widget_ids;
 
 pub use nalgebra;
 pub use ncollide3d;
