+++ src/renderer/mod.rs
@@ -1,13 +1,9 @@
 //! Structures responsible for rendering elements other than kiss3d's meshes.
 
-#[cfg(feature = "conrod")]
-pub use self::conrod_renderer::ConrodRenderer;
 pub use self::line_renderer::LineRenderer;
 pub use self::point_renderer::PointRenderer;
 pub use self::renderer::Renderer;
 
-#[cfg(feature = "conrod")]
-mod conrod_renderer;
 pub mod line_renderer;
 pub mod point_renderer;
 mod renderer;
