+++ src/window/mod.rs
@@ -7,7 +7,7 @@ pub use self::gl_canvas::GLCanvas;
 pub use self::state::State;
 #[cfg(target_arch = "wasm32")]
 pub use self::webgl_canvas::WebGLCanvas;
-pub use self::window::Window;
+pub use self::window::{UiContext, Window};
 
 mod canvas;
 #[cfg(not(target_arch = "wasm32"))]
