+++ ui/conrod/examples/ui_conrod.rs
@@ -1,27 +1,13 @@
 extern crate kiss3d;
 extern crate nalgebra as na;
 
-#[cfg(feature = "conrod")]
-use kiss3d::conrod;
-#[cfg(feature = "conrod")]
-use kiss3d::conrod::color::Color;
-#[cfg(feature = "conrod")]
-use kiss3d::conrod::position::Positionable;
-#[cfg(feature = "conrod")]
-use kiss3d::conrod::widget_ids;
-#[cfg(feature = "conrod")]
+use conrod::position::Positionable;
+use conrod::widget_ids;
 use kiss3d::light::Light;
-#[cfg(feature = "conrod")]
 use kiss3d::window::Window;
-#[cfg(feature = "conrod")]
+use kiss3d_ui_conrod::{conrod, ConrodContext};
 use std::path::Path;
 
-#[cfg(not(feature = "conrod"))]
-fn main() {
-    panic!("The 'conrod' feature must be enabled for this example to work.")
-}
-
-#[cfg(feature = "conrod")]
 fn main() {
     let mut window = Window::new("Kiss3d: UI");
     window.set_background_color(1.0, 1.0, 1.0);
@@ -30,17 +16,23 @@ fn main() {
 
     window.set_light(Light::StickToCamera);
 
+    // Initialize Conrod UI.
+    let mut conrod_ctx = Box::new(ConrodContext::new(window.width(), window.height()));
+
     // Generate the widget identifiers.
-    let ids = Ids::new(window.conrod_ui_mut().widget_id_generator());
-    window.conrod_ui_mut().theme = theme();
+    let ids = Ids::new(conrod_ctx.conrod_ui_mut().widget_id_generator());
+    conrod_ctx.conrod_ui_mut().theme = theme();
     window.add_texture(&Path::new("./examples/media/kitten.png"), "cat");
-    let cat_texture = window.conrod_texture_id("cat").unwrap();
+    let cat_texture = conrod_ctx.conrod_texture_id("cat").unwrap();
+
+    window.set_ui(conrod_ctx);
 
     let mut app = DemoApp::new(cat_texture);
 
     // Render loop.
     while window.render() {
-        let mut ui = window.conrod_ui_mut().set_widgets();
+        let conrod_ctx = window.ui_mut::<ConrodContext>().unwrap();
+        let mut ui = conrod_ctx.conrod_ui_mut().set_widgets();
         gui(&mut ui, &ids, &mut app)
     }
 }
@@ -51,7 +43,6 @@ fn main() {
  *
  */
 /// A set of reasonable stylistic defaults that works for the `gui` below.
-#[cfg(feature = "conrod")]
 pub fn theme() -> conrod::Theme {
     use conrod::position::{Align, Direction, Padding, Position, Relative};
     conrod::Theme {
@@ -75,7 +66,6 @@ pub fn theme() -> conrod::Theme {
 }
 
 // Generate a unique `WidgetId` for each widget.
-#[cfg(feature = "conrod")]
 widget_ids! {
     pub struct Ids {
         // The scrollable canvas.
@@ -122,7 +112,6 @@ pub const WIN_W: u32 = 600;
 pub const WIN_H: u32 = 420;
 
 /// A demonstration of some application state we want to control with a conrod GUI.
-#[cfg(feature = "conrod")]
 pub struct DemoApp {
     ball_xy: conrod::Point,
     ball_color: conrod::Color,
@@ -132,7 +121,6 @@ pub struct DemoApp {
     text_edit: String,
 }
 
-#[cfg(feature = "conrod")]
 impl DemoApp {
     /// Simple constructor for the `DemoApp`.
     pub fn new(cat: conrod::image::Id) -> Self {
@@ -148,7 +136,6 @@ impl DemoApp {
 }
 
 /// Instantiate a GUI demonstrating every widget available in conrod.
-#[cfg(feature = "conrod")]
 pub fn gui(ui: &mut conrod::UiCell, ids: &Ids, app: &mut DemoApp) {
     use conrod::{widget, Colorable, Labelable, Sizeable, Widget};
     use std::iter::once;
